+++ spec/rubocop/cop/lint/memoization_with_parameters_spec.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-RSpec.describe RuboCop::Cop::Lint::MemorizationWithParameters do
+RSpec.describe RuboCop::Cop::Lint::MemoizationWithParameters do
   subject(:cop) { described_class.new(config) }
 
   let(:config) { RuboCop::Config.new }
   let(:error_message) do
-    'Use an instance variable such as a Hash to store method returns called ' \
-    'with parameter(s).'
+    'Use an instance variable such as a Hash to store the ' \
+    'result of a computation depending on parameter(s).'
   end
 
   before do
@@ -97,6 +97,33 @@ RSpec.describe RuboCop::Cop::Lint::MemorizationWithParameters do
           @foo ||= begin
             bar
           end
+        end
+      RUBY
+    end
+
+    it { expect(cop.offenses.size).to eq(1) }
+    it { expect(cop.offenses.first.message).to eq error_message }
+  end
+
+  context 'when method contains more than just memoization' do
+    subject(:source) do
+      <<-'RUBY'.strip_indent
+        def foo(bar)
+          puts "Called with #{bar}"
+          @foo ||= baz(bar)
+        end
+      RUBY
+    end
+
+    it { expect(cop.offenses.size).to eq(1) }
+    it { expect(cop.offenses.first.message).to eq error_message }
+  end
+
+  context 'with multiple method calls' do
+    subject(:source) do
+      <<-RUBY.strip_indent
+        def foo(bar)
+          @foo ||= qux(baz(bar))
         end
       RUBY
     end